requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.2
django-allauth==0.44.0
django-appconf==1.0.4
django-class-auth-views==1.0.2
django-classy-tags==2.0.0
django-cms==3.8.0
django-crm==0.8.0
django-formtools==2.3
django-ipware==3.0.2
django-js-asset==1.2.2
django-mptt==0.12.0
django-sekizai==2.0.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
gunicorn==20.1.0
heroku==0.1.4
idna==2.10
mysql==0.0.2
mysqlclient==2.0.3
oauthlib==3.1.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==1.5
python3-openid==3.2.0
pytz==2021.1
razorpay==1.2.0
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
